Spare 8 seconds on QT100 decoding
Realized the dcraw algorithm, from which it's derived, does things in an
"easy, cycles are free" way in more places than I had initially seen.
There's a whole "init the full buffer to grey" thing and then these grey
values are reused later. But in fact it's only necessary to init the two
first lines with grey, and then in the second loop I can now hardcode an
lda #$80 instead of lda (idx),y...  and then I realize that this #$80 is
left-shifted once, and later we substract... $100... and so this math is
100% useless.