Rework draw functions for better results
and less RAM usage. Go back to each sprite saving its background
itself, get rid of the 8kB copy of the background, and order
sprite clear/draw so that they don't keep parts of each other
when clearing.
Rework draw order according to the puck's side of the table so
that there is less flickering